Made a bunch of more non-deprecated casts.

+++ src/containers/buffer_group.cc
@@ -3,12 +3,13 @@
 
 void buffer_group_copy_data(const buffer_group_t *dest, const const_buffer_group_t *source) {
 
-    int bytes = source->get_size();
-    rassert(bytes == (int)dest->get_size());
+    // TODO: Is a buffer group size an int?
+    int64_t bytes = source->get_size();
+    rassert(bytes == static_cast<int64_t>(dest->get_size()));
 
     /* Copy data between source and dest; we have to always copy the minimum of the sizes of the
     next chunk that each one has */
-    int source_buf = 0, source_off = 0, dest_buf = 0, dest_off = 0;
+    int64_t source_buf = 0, source_off = 0, dest_buf = 0, dest_off = 0;
     while (bytes > 0) {
         while (source->get_buffer(source_buf).size == source_off) {
             ++source_buf;
@@ -31,12 +32,10 @@ void buffer_group_copy_data(const buffer_group_t *dest, const const_buffer_group
     }
 
     /* Make sure we reached the end of both source and dest */
-    rassert(
-        (source_buf == (int)source->num_buffers()     && source_off == 0) ||
-        (source_buf == (int)source->num_buffers() - 1 && source_off == source->get_buffer(source_buf).size));
-    rassert(
-        (dest_buf == (int)dest->num_buffers()     && dest_off == 0) ||
-        (dest_buf == (int)dest->num_buffers() - 1 && dest_off == dest->get_buffer(dest_buf).size));
+    rassert((source_buf == static_cast<int64_t>(source->num_buffers())     && source_off == 0) ||
+            (source_buf == static_cast<int64_t>(source->num_buffers()) - 1 && source_off == source->get_buffer(source_buf).size));
+    rassert((dest_buf == static_cast<int64_t>(dest->num_buffers())     && dest_off == 0) ||
+            (dest_buf == static_cast<int64_t>(dest->num_buffers()) - 1 && dest_off == dest->get_buffer(dest_buf).size));
 }
 
 void buffer_group_copy_data(const buffer_group_t *out, const char *in, int64_t size) {
